(refactoring) move onNewBlock to top level;
