+++ client/src/app/routes/stardust/VisualizerContainer.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { RouteComponentProps } from "react-router-dom";
-import VisualizerThree from "../../../features/visualizer-threejs/VisualizerThree";
+import VisualizerThree from "../../../features/visualizer-threejs/VisualizerMain";
 import { VisualizerRouteProps } from "../VisualizerRouteProps";
 import { VisualizerDefault } from "./VisualizerDefault";
 
@@ -11,7 +11,7 @@ enum Views {
 
 export const VisualizerContainer: React.FC<
     RouteComponentProps<VisualizerRouteProps>
-> = (props) => {
+> = props => {
     const [currentView] = useState(Views.three);
 
     if (currentView === Views.default) {
