+++ client/src/features/visualizer-threejs/hooks/useZoomDynamic.ts
@@ -1,0 +1,19 @@
+import { useThree } from "@react-three/fiber";
+import { useEffect } from "react";
+import { useBlockStore } from "../store";
+
+export const useZoomDynamic = () => {
+    /**
+     * Zoom Changes
+     */
+    const cameraState = useThree(state => state.camera);
+    const zoomStore = useBlockStore(s => s.zoom);
+    useEffect(() => {
+        if (cameraState) {
+            cameraState.zoom = zoomStore;
+            cameraState.updateProjectionMatrix();
+        }
+    }, [cameraState, zoomStore]);
+
+    return {};
+};
