+++ client/src/features/visualizer-threejs/types.ts
@@ -1,5 +1,5 @@
-import { IFeedBlockData } from "../../../models/api/stardust/feed/IFeedBlockData";
-import { IFeedBlockMetadata } from "../../../models/api/stardust/feed/IFeedBlockMetadata";
+import { IFeedBlockData } from "../../models/api/stardust/feed/IFeedBlockData";
+import { IFeedBlockMetadata } from "../../models/api/stardust/feed/IFeedBlockMetadata";
 
 export type TFeedBlockAdd = (newBlock: IFeedBlockData) => void;
