+++ client/src/features/visualizer-threejs/worker/entities/DataSender.ts
@@ -1,4 +1,4 @@
-import { DATA_SENDER_TIME_INTERVAL } from "../../../../shared/visualizer/common/constants";
+import { DATA_SENDER_TIME_INTERVAL } from "../../constants";
 
 export class DataSender {
     private lastSentTimestamp: number = 0;
