+++ client/src/features/visualizer-webgl/layout.ts
@@ -5,8 +5,30 @@
 /* eslint-disable @typescript-eslint/no-invalid-this */
 import Viva from "vivagraphjs";
 import { INodeData } from "../../models/graph/stardust/INodeData";
-import { generateCoordinateGrid } from "../../shared/visualizer/common/utils";
 import { VivaLink } from "./vivagraph-layout.types";
+
+/**
+ * Function for generating list of coordinates
+ * @param start - start of coordinates. Example: -300
+ * @param end - end of coordinates. Example - 300
+ * @param numberOfPoints - how much points you need between start and end.
+ * @returns list of coordinates
+ */
+export function generateCoordinateGrid(
+    start: number,
+    end: number,
+    numberOfPoints: number
+): number[] {
+    const coordinates: number[] = [];
+
+    const increment = (end - start) / (numberOfPoints - 1);
+
+    for (let i = 0, y = start; i < numberOfPoints; i++, y += increment) {
+        coordinates.push(y);
+    }
+
+    return coordinates;
+}
 
 class Rect {
     public x1: number;
